requirements.txt
yt-dlp
opencv-python
openai
pillow
requests
fastapi
uvicorn
